Add handleColor and handleBlock in markedContex
